+++ tests/sso_token_util_tests.c
@@ -54,7 +54,7 @@ static int s_parse_sso_token_valid(struct aws_allocator *allocator, void *ctx) {
     ASSERT_TRUE(aws_create_profile_file(file_path, s_valid_token_json) == AWS_OP_SUCCESS);
     struct aws_sso_token *token = aws_sso_token_new_from_file(allocator, file_path);
     ASSERT_TRUE(aws_string_eq_c_str(token->token, "string"));
-    ASSERT_INT_EQUALS(aws_date_time_as_epoch_secs(&token->expiration), 1573704345);
+    ASSERT_INT_EQUALS((uint64_t)aws_date_time_as_epoch_secs(&token->expiration), 1573704345);
     aws_string_destroy(file_path);
     aws_sso_token_destroy(token);
     aws_auth_library_clean_up();
